moved PipetteName location
