+++ api/src/opentrons/protocol_engine/__init__.py
@@ -33,7 +33,6 @@ from .types import (
     LoadedLabware,
     LoadedModule,
     LoadedPipette,
-    PipetteName,
     WellLocation,
     WellOrigin,
     WellOffset,
@@ -78,7 +77,6 @@ __all__ = [
     "LoadedLabware",
     "LoadedModule",
     "LoadedPipette",
-    "PipetteName",
     "WellLocation",
     "WellOrigin",
     "WellOffset",
