+++ api/tests/opentrons/protocol_engine/commands/test_load_pipette.py
@@ -1,8 +1,7 @@
 """Test load pipette commands."""
 from decoy import Decoy
 
-from opentrons.types import MountType
-from opentrons.protocol_engine.types import PipetteName
+from opentrons.types import MountType, PipetteName
 from opentrons.protocol_engine.execution import LoadedPipetteData, EquipmentHandler
 
 from opentrons.protocol_engine.commands.load_pipette import (
