+++ api/tests/opentrons/protocol_engine/state/command_fixtures.py
@@ -3,11 +3,10 @@ from datetime import datetime
 from pydantic import BaseModel
 from typing import Optional, cast
 
-from opentrons.types import MountType
+from opentrons.types import MountType, PipetteName
 from opentrons.protocols.models import LabwareDefinition
 from opentrons.protocol_engine import ErrorOccurrence, commands as cmd
 from opentrons.protocol_engine.types import (
-    PipetteName,
     WellLocation,
     LabwareLocation,
 )
